mtd-utils: nanddump: Allow 64-bit lengths
We should allow the dump length to be 64-bit, especially since the value
was read in as a "long long" by strtoll().

Acked-by: Mike Frysinger <vapier@gentoo.org>

+++ nanddump.c
@@ -83,7 +83,7 @@ static bool			noecc = false;		// don't error correct
 static bool			noskipbad = false;	// don't skip bad blocks
 static bool			omitoob = false;	// omit oob data
 static unsigned long long	start_addr;		// start address
-static unsigned long		length;			// dump length
+static unsigned long long	length;			// dump length
 static const char		*mtddev;		// mtd device name
 static const char		*dumpfile;		// dump file name
 static bool			omitbad = false;
